+++ NEWS.md
@@ -4,6 +4,7 @@
   arguments to contain objects (e.g. as might happen when a function is invoked
   with `do.call`).
 * Address fun(void) declarations.
+* Swap `snprintf` for `sprintf` (and fix off by one allocations).
 
 ## 0.2.14